Move json jsoniter unmarshal to be reused for potlp packages
Fix https://github.com/open-telemetry/opentelemetry-collector/issues/6027
