+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 
 - Add `skip-get-modules` builder flag to support isolated environment executions (#6009)
   - Skip unnecessary Go binary path validation when the builder is used with `skip-compilation` and `skip-get-modules` flags (#6026)
+- Make the otlpreceiver support to use jsoniter to unmarshal JSON payloads. (#6040)
 
 ### 🧰 Bug fixes 🧰
